refactor(ui): isolate project prompt preparation
