+++ packages/ui/src/components/thread-playground/codegen/generate-project-button.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import {
-  getMessageText,
   type McpServerView,
   type ModelConfig,
   type ModelProviderGroup,
@@ -19,7 +18,6 @@ import {
   type GeneratorModelInfo,
   type GeneratorResult,
 } from "@llm-space/core/generator";
-import { renderThreadPromptVariables } from "@llm-space/core/thread";
 import {
   createOneShotRunner,
   createWorkflowContext,
@@ -43,7 +41,6 @@ import { useHostServices } from "@llm-space/ui/host";
 import { Spinner } from "@llm-space/ui/ui/spinner";
 import { Switch } from "@llm-space/ui/ui/switch";
 
-import type { FilesHost } from "../../../host/types";
 import { cn } from "../../../lib/utils";
 import { Button } from "../../../ui/button";
 import {
@@ -58,13 +55,13 @@ import { Input } from "../../../ui/input";
 import { ConfirmDialog } from "../../confirm-dialog";
 import { useFirstAvailableModel, useModels } from "../../model-provider";
 import { Tooltip } from "../../tooltip";
-import { createRuntimePromptFiles } from "../runtime-prompt-files";
 import {
   useThreadStore,
   useThreadStoreApi,
-  type ThreadStore,
 } from "../stores/thread-store";
 import { listEnabledPromptVariableSkills } from "../variable/prompt-variable-skills";
+
+import { createGenerateProjectPromptPreparer } from "./generate-project-prompt-preparer";
 
 /** The generator to run. V1 ships only LangGraph. */
 const GENERATOR_ID = "langgraph";
@@ -89,81 +86,6 @@ const FRAMEWORKS = [
 ] as const;
 
 type WizardStep = "framework" | "target" | "run";
-
-/**
- * Prepare the prompt inputs consumed by the real Generate Project workflow.
- * Runtime ownership is explicit so this path cannot silently read same-named
- * files from the desktop's local runtime.
- */
-async function _prepareGenerateProjectPromptContext({
-  context,
-  files,
-  runtimeId,
-  skillList,
-  useMetaUserPrompt,
-}: {
-  context: ThreadContext;
-  files: FilesHost;
-  runtimeId: string;
-  skillList: Awaited<ReturnType<typeof listEnabledPromptVariableSkills>>;
-  useMetaUserPrompt: boolean;
-}) {
-  const promptFiles = createRuntimePromptFiles(files, runtimeId);
-  const rendered = await renderThreadPromptVariables({
-    context,
-    loadSkills: () => Promise.resolve(skillList),
-    loadFile: promptFiles.loadFile,
-    fileExists: promptFiles.fileExists,
-  });
-  const systemPromptTemplate = await _expandIncludes(
-    context.systemPrompt ?? "",
-    promptFiles.loadFile
-  );
-  const firstMessage = context.messages?.[0];
-  const firstUserMessageTemplate =
-    useMetaUserPrompt && firstMessage?.role === "user"
-      ? await _expandIncludes(getMessageText(firstMessage), promptFiles.loadFile)
-      : undefined;
-  const renderedVariableValues: Record<string, string> = Object.fromEntries(
-    rendered.variables.map((variable) => [variable.name, variable.value])
-  );
-
-  return {
-    rendered,
-    systemPromptTemplate,
-    firstUserMessageTemplate,
-    renderedVariableValues,
-  };
-}
-
-/** Bind Generate Project prompt reads directly to one owning thread store. */
-export function createGenerateProjectPromptPreparer({
-  files,
-  store,
-}: {
-  files: FilesHost;
-  store: ThreadStore;
-}) {
-  return ({
-    skillList,
-    useMetaUserPrompt,
-  }: {
-    skillList: Awaited<ReturnType<typeof listEnabledPromptVariableSkills>>;
-    useMetaUserPrompt: boolean;
-  }) => {
-    const { runtimeId, thread } = store.getState();
-    if (!runtimeId) {
-      throw new Error("Generate Project requires an owning runtimeId");
-    }
-    return _prepareGenerateProjectPromptContext({
-      context: thread.context ?? {},
-      files,
-      runtimeId,
-      skillList,
-      useMetaUserPrompt,
-    });
-  };
-}
 
 /**
  * Header action that exports the current thread as a runnable code project via
@@ -1230,42 +1152,6 @@ function _shellQuote(path: string): string {
   return /^[A-Za-z0-9_./~-]+$/.test(path)
     ? path
     : `'${path.replace(/'/g, "'\\''")}'`;
-}
-
-/** `{{@include("path")}}` macro with a single quoted path argument. */
-const INCLUDE_MACRO_RE = /\{\{\s*@include\(\s*(['"])([\s\S]*?)\1\s*\)\s*\}\}/g;
-
-/**
- * Expand LLM Space's `@include(...)` macro by inlining the referenced files,
- * so the generated project (which renders with plain Jinja2) doesn't need it.
- * Bounded recursion guards against include cycles.
- */
-async function _expandIncludes(
-  text: string,
-  readText: (path: string) => Promise<string>,
-  depth = 0
-): Promise<string> {
-  if (depth > 10 || !text.includes("@include")) {
-    return text;
-  }
-  const matches = [...text.matchAll(INCLUDE_MACRO_RE)];
-  if (matches.length === 0) {
-    return text;
-  }
-  let out = "";
-  let last = 0;
-  for (const match of matches) {
-    out += text.slice(last, match.index);
-    let content: string;
-    try {
-      content = await readText(match[2]);
-    } catch {
-      content = "";
-    }
-    out += await _expandIncludes(content, readText, depth + 1);
-    last = (match.index ?? 0) + match[0].length;
-  }
-  return out + text.slice(last);
 }
 
 function ProgressLine({ event }: { event: WorkflowEvent }) {
